報酬数を調整できるように修正

+++ domain/service/example/example_service_mock.gen.go
@@ -35,7 +35,7 @@ func (m *MockExampleService) EXPECT() *MockExampleServiceMockRecorder {
 }
 
 // ListExample mocks base method.
-func (m *MockExampleService) ListExample(limit int64) (*example.ListExample, error) {
+func (m *MockExampleService) ListExample(limit int) (*example.ListExample, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ListExample", limit)
 	ret0, _ := ret[0].(*example.ListExample)
